integrated new matching classes in development branch
